+++ packages/agent-infra/mcp-servers/browser/src/server.ts
@@ -16,6 +16,7 @@ import {
   ImageContent,
   TextContent,
 } from '@modelcontextprotocol/sdk/types.js';
+import { toMarkdown } from '@agent-infra/shared';
 import { Logger, ConsoleLogger } from '@agent-infra/logger';
 import { z } from 'zod';
 import {
@@ -38,12 +39,10 @@ import {
   locateElement,
   scrollIntoViewIfNeeded,
 } from '@agent-infra/browser-use';
-import TurndownService from 'turndown';
-// @ts-ignore
-import { gfm } from 'turndown-plugin-gfm';
 import merge from 'lodash.merge';
 import { parseProxyUrl } from './utils.js';
-import { ElementHandle } from 'puppeteer-core';
+import { ElementHandle, KeyInput } from 'puppeteer-core';
+import { keyInputValues } from './constants.js';
 
 const consoleLogs: string[] = [];
 
@@ -236,11 +235,14 @@ export const toolsMap = {
           .number()
           .optional()
           .describe('index of the element to screenshot'),
-        width: z.number().optional().describe('Width in pixels (default: 800)'),
+        width: z
+          .number()
+          .optional()
+          .describe('Width in pixels (default: viewport width)'),
         height: z
           .number()
           .optional()
-          .describe('Height in pixels (default: 600)'),
+          .describe('Height in pixels (default: viewport height)'),
         highlight: z
           .boolean()
           .optional()
@@ -268,7 +270,7 @@ export const toolsMap = {
   },
   browser_form_input_fill: {
     name: 'browser_form_input_fill',
-    description: 'Fill out an input field',
+    description: 'Fill out an input field, before using the tool',
     inputSchema: z
       .object({
         selector: z
@@ -323,7 +325,8 @@ export const toolsMap = {
   // new tools
   browser_get_html: {
     name: 'browser_get_html',
-    description: 'Get the HTML content of the current page',
+    description:
+      'Get the HTML content of the current page, return long text is not friendly to models with limited token, recommended for use browser_get_markdown instead',
   },
   browser_get_clickable_elements: {
     name: 'browser_get_clickable_elements',
@@ -348,7 +351,10 @@ export const toolsMap = {
     inputSchema: z.object({
       amount: z
         .number()
-        .describe('Pixels to scroll (positive for down, negative for up)'),
+        .optional()
+        .describe(
+          'Pixels to scroll (positive for down, negative for up), if the amount is not provided, scroll to the bottom of the page',
+        ),
     }),
   },
   browser_go_back: {
@@ -379,6 +385,19 @@ export const toolsMap = {
     description: 'Switch to a specific tab',
     inputSchema: z.object({
       index: z.number().describe('Tab index to switch to'),
+    }),
+  },
+  browser_press_key: {
+    name: 'browser_press_key',
+    description: 'Press a key on the keyboard',
+    inputSchema: z.object({
+      key: z
+        .enum(keyInputValues as [string, ...string[]])
+        .describe(
+          `Name of the key to press or a character to generate, such as ${keyInputValues.join(
+            ', ',
+          )}`,
+        ),
     }),
   },
 };
@@ -534,11 +553,12 @@ const handleToolCall = async ({
         ]);
         logger.info('navigateTo complete');
         const { clickableElements } = (await buildDomTree(page)) || {};
+        await removeHighlights(page);
         return {
           content: [
             {
               type: 'text',
-              text: `Navigated to ${args.url}\nclickable elements: ${clickableElements}`,
+              text: `Navigated to ${args.url}\nclickable elements(Might be outdated, if an error occurs with the index element, use browser_get_clickable_elements to refresh it): ${clickableElements}`,
             },
           ],
           isError: false,
@@ -1001,17 +1021,8 @@ const handleToolCall = async ({
     },
     browser_get_markdown: async (args) => {
       try {
-        const turndownService = new TurndownService();
-        turndownService.addRule('filter_tags', {
-          filter: ['script', 'style'],
-          replacement: (content) => {
-            return '';
-          },
-        });
-        turndownService.use(gfm);
-
         const html = await page.content();
-        const markdown = turndownService.turndown(html);
+        const markdown = toMarkdown(html);
         return {
           content: [{ type: 'text', text: markdown }],
           isError: false,
@@ -1211,6 +1222,20 @@ const handleToolCall = async ({
               text: `Failed to switch tab: ${(error as Error).message}`,
             },
           ],
+          isError: true,
+        };
+      }
+    },
+    browser_press_key: async (args) => {
+      try {
+        await page.keyboard.press(args.key as KeyInput);
+        return {
+          content: [{ type: 'text', text: `Pressed key: ${args.key}` }],
+          isError: false,
+        };
+      } catch (error) {
+        return {
+          content: [{ type: 'text', text: `Failed to press key: ${args.key}` }],
           isError: true,
         };
       }