+++ packages/agent-infra/mcp-servers/filesystem/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @agent-infra/mcp-server-filesystem
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.4
+
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.3
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes